requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-extensions==2.1.6
django-summernote==0.8.11.4
djangorestframework==3.9.2
mysqlclient==1.3.13
Pillow==5.4.1
pygraphviz==1.5
pytz==2018.9
six==1.12.0
